moved visualisation to d_graphs
